1.Appling hall input linier pwm output

+++ Projects/N76E003_BSP_Keil_C51_V1.0.6/Include/Motor_control.h
@@ -1,6 +1,6 @@
 #define u8 unsigned char
 #define u16 unsigned int
 void ADC_Init(void);
-u8 Get_HallValue(void);
+u16 Get_HallValue(void);
 void PWM_Init(void);
-void PWM_Setting(double n);
+void PWM_Setting(u16 n);
